UI is now in color.

+++ vbox-monitor/VBox/String.hpp
@@ -62,7 +62,7 @@ namespace VBox
                << std::uppercase
                << std::setfill( '0' )
                << std::setw( sizeof( _T_ ) * 2 )
-               << v;
+               << static_cast< uint64_t >( v );
             
             return ss.str();
         }
